# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 建行杯小程序 API
  version: 0.1.0
  description: |
    纸韵蛙鸣·哈什蚂传奇小程序第一阶段接口契约。
    成功响应 code 固定为字符串 "0"；失败响应使用模块化错误码。
servers:
  - url: http://localhost:8080
    description: 本地开发
  - url: https://api.example.com
    description: 生产环境占位，部署后替换
tags:
  - name: 认证
  - name: 首页
  - name: 用户
  - name: 游戏
  - name: 解锁
  - name: 排行榜
  - name: 公益
  - name: 内容
  - name: 系统
paths:
  /v1/auth/wechat-login:
    post:
      tags: [认证]
      summary: 微信小程序登录
      operationId: wechatLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /v1/home/summary:
    get:
      tags: [首页]
      summary: 获取首页聚合数据
      operationId: getHomeSummary
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/content/frogs:
    get:
      tags: [内容]
      summary: 获取林蛙角色与素材清单
      operationId: getFrogAssets
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrogAssetResponse'
  /v1/content/frogs/{frogId}:
    get:
      tags: [内容]
      summary: 获取单只林蛙详情
      operationId: getFrogDetail
      parameters:
        - name: frogId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FrogDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/me/profile:
    get:
      tags: [用户]
      summary: 获取当前用户资料
      operationId: getMyProfile
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [用户]
      summary: 更新当前用户资料
      operationId: updateMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/games/{gameId}/progress:
    get:
      tags: [游戏]
      summary: 获取游戏进度
      operationId: getGameProgress
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameProgressResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/games/{gameId}/events:
    post:
      tags: [游戏]
      summary: 提交游戏事件
      description: 服务端校验游戏事件并按幂等键去重。
      operationId: submitGameEvent
      parameters:
        - $ref: '#/components/parameters/GameId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameEventRequest'
      responses:
        '200':
          description: 事件已接收
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameEventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/unlocks/redeem:
    post:
      tags: [解锁]
      summary: 核销扫码解锁码
      operationId: redeemUnlockCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockRedeemRequest'
      responses:
        '200':
          description: 核销成功或返回已核销结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockRedeemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/rankings:
    get:
      tags: [排行榜]
      summary: 获取排行榜
      operationId: getRankings
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [total, weekly]
            default: total
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/welfare/summary:
    get:
      tags: [公益]
      summary: 获取公益基金摘要
      operationId: getWelfareSummary
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WelfareSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/system/ping:
    get:
      tags: [系统]
      summary: 服务连通性检查
      operationId: ping
      security: []
      responses:
        '200':
          description: 服务可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
        enum: [story, guardian, forest-quiz]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 每次业务操作生成的新键；重试同一操作时复用原键。服务端不强制要求，
        但缺失时无法做重放去重，客户端应始终携带。
      schema:
        type: string
        maxLength: 64
  responses:
    BadRequest:
      description: 请求参数不正确
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未登录或登录状态失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源状态冲突或幂等键内容不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UpstreamError:
      description: 微信上游服务异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ResponseMeta:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          enum: ['0']
        message:
          type: string
          example: success
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
          nullable: true
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          example: COMMON_400
          enum: [COMMON_400, AUTH_401, AUTH_403, COMMON_404, COMMON_409, COMMON_500, WECHAT_502]
        message:
          type: string
          example: 请求参数不正确
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
          nullable: true
    WechatLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 128
          description: wx.login 返回的临时登录凭证
        nickname:
          type: string
          maxLength: 64
        avatarUrl:
          type: string
          format: uri
          maxLength: 512
    TokenData:
      type: object
      required: [accessToken, expiresIn, user]
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          example: 7200
        user:
          $ref: '#/components/schemas/User'
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/TokenData'
    User:
      type: object
      required: [id, nickname, avatarUrl]
      properties:
        id:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
    GameCard:
      type: object
      required: [id, title, subtitle, path]
      properties:
        id:
          type: string
          enum: [paper-cutting, story, guardian, forest-quiz]
        title:
          type: string
        subtitle:
          type: string
        path:
          type: string
    Frog:
      type: object
      required: [id, name, shortName, pattern, blessing, assetUrl, unlocked]
      properties:
        id:
          type: string
        name:
          type: string
        shortName:
          type: string
        pattern:
          type: string
        blessing:
          type: string
        assetUrl:
          type: string
          format: uri-reference
        unlocked:
          type: boolean
    FrogAssetResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Frog'
        total:
          type: integer
          minimum: 0
    FrogDetailSection:
      type: object
      required: [heading, paragraphs]
      properties:
        heading:
          type: string
        paragraphs:
          type: array
          items:
            type: string
    FrogDetailResponse:
      allOf:
        - $ref: '#/components/schemas/Frog'
        - type: object
          required: [sections]
          properties:
            sections:
              type: array
              items:
                $ref: '#/components/schemas/FrogDetailSection'
    HomeSummary:
      type: object
      required: [brand, slogan, fundAmount, games, frogs, activity]
      properties:
        brand:
          type: string
        slogan:
          type: string
        fundAmount:
          type: string
          example: '12,480.00'
        games:
          type: array
          items:
            $ref: '#/components/schemas/GameCard'
        frogs:
          type: array
          items:
            $ref: '#/components/schemas/Frog'
        activity:
          type: array
          description: 只包含真实的徽章解锁流水，没有解锁时为空数组。
          items:
            type: string
    HomeSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/HomeSummary'
    ProfileStats:
      type: object
      required: [frogs, frogsTotal, badges, badgesTotal]
      properties:
        frogs: { type: integer, minimum: 0 }
        frogsTotal: { type: integer, minimum: 0 }
        badges: { type: integer, minimum: 0 }
        badgesTotal: { type: integer, minimum: 0 }
    Badge:
      type: object
      required: [id, name, level, unlocked]
      properties:
        id: { type: string }
        name: { type: string }
        level:
          type: string
          enum: [bronze, silver, gold]
        unlocked: { type: boolean }
    ProfileData:
      type: object
      required: [user, stats, badges]
      properties:
        user: { $ref: '#/components/schemas/User' }
        stats: { $ref: '#/components/schemas/ProfileStats' }
        badges:
          type: array
          items: { $ref: '#/components/schemas/Badge' }
    ProfileResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ProfileData' }
    ProfileUpdateRequest:
      type: object
      minProperties: 1
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 64
        avatarUrl:
          type: string
          format: uri
          maxLength: 512
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/User' }
    GameProgress:
      type: object
      required: [gameId, completed, total, finished, version, updatedAt]
      properties:
        gameId:
          type: string
          enum: [story, guardian, forest-quiz]
        completed:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 1
        finished:
          type: boolean
        version:
          type: integer
          minimum: 0
        state:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
    GameProgressResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/GameProgress' }
    GameEventRequest:
      type: object
      required: [type]
      additionalProperties: true
      properties:
        type:
          type: string
          enum: [story_choice, reset, draw, answer, share]
        payload:
          type: object
          additionalProperties: true
          description: '故事事件使用 choiceId，例如 { "choiceId": "A" }；守护神 answer 使用 roundId 与 pattern，例如 { "roundId": "uuid", "pattern": "蛙纹" }。draw 和 share 不需要 payload。'
        clientOccurredAt:
          type: string
          format: date-time
    GameEventResult:
      type: object
      required: [accepted, duplicated, gameId, progress]
      properties:
        accepted: { type: boolean }
        duplicated: { type: boolean }
        gameId: { type: string }
        progress: { $ref: '#/components/schemas/GameProgress' }
        action:
          type: object
          additionalProperties: true
          description: '守护神事件结果。draw 返回 drawn 与题目提示；answer 返回 correct、guardianCard 与 newlyUnlockedBadges；share 返回 shareBonusGranted。guardian 进度 state 包含当天日期、剩余次数、分享领取状态、图鉴、当前题目和徽章。'
    GameEventResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/GameEventResult' }
    UnlockRedeemRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 64
    UnlockRedeemResult:
      type: object
      required: [accepted, code, characterId, characterName, message, redeemedAt]
      properties:
        accepted: { type: boolean }
        code: { type: string }
        characterId: { type: string }
        characterName: { type: string }
        message: { type: string }
        redeemedAt: { type: string, format: date-time }
    UnlockRedeemResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/UnlockRedeemResult' }
    RankingItem:
      type: object
      required: [rank, userId, nickname, score]
      properties:
        rank: { type: integer, minimum: 1 }
        userId: { type: string }
        nickname: { type: string }
        score: { type: integer, minimum: 0 }
    RankingData:
      type: object
      required: [type, updatedAt, items, page, size, total]
      properties:
        type: { type: string, enum: [total, weekly] }
        updatedAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/RankingItem' }
        myRank:
          type: integer
          minimum: 1
          nullable: true
          description: 当前用户的名次；没有徽章的用户不上榜，字段被省略。
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
    RankingResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/RankingData' }
    WelfareReport:
      type: object
      required: [id, title, period, amount, publishedAt]
      properties:
        id: { type: string }
        title: { type: string }
        period: { type: string }
        amount: { type: string }
        publishedAt: { type: string, format: date-time }
        documentUrl: { type: string, format: uri }
    WelfareSummary:
      type: object
      required: [fundAmount, reports, updatedAt]
      properties:
        fundAmount: { type: string, example: '12,480.00' }
        reports:
          type: array
          items: { $ref: '#/components/schemas/WelfareReport' }
        updatedAt: { type: string, format: date-time }
    WelfareSummaryResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/WelfareSummary' }
    PingData:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [up] }
    PingResponse:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/PingData' }
security:
  - bearerAuth: []